Internship-1/week_9_training_search_engine_with_lemmatization_tfidf_and_cosine_similarity/search_system/requirements.txt
# needed dependecies

Flask==3.1.0
Flask-Cors==5.0.0
nltk==3.9.1
scikit-learn==1.6.1
requests==2.32.3
beautifulsoup4==4.12.3
python-docx==1.1.2
PyPDF2==3.0.1


# to replicate full venv

# asgiref==3.8.1
# blinker==1.9.0
# certifi==2024.12.14
# cffi==1.17.1
# charset-normalizer==3.4.1
# click==8.1.8
# cryptography==44.0.0
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.5
# joblib==1.4.2
# lxml==5.3.0
# MarkupSafe==3.0.2
# numpy==2.2.2
# pdfminer.six==20231228
# pdfplumber==0.11.5
# pillow==11.1.0
# pycparser==2.22
# pypdfium2==4.30.1
# regex==2024.11.6
# scipy==1.15.1
# soupsieve==2.6
# sqlparse==0.5.3
# threadpoolctl==3.5.0
# tqdm==4.67.1
# typing_extensions==4.12.2
# urllib3==2.3.0
# Werkzeug==3.1.3
